+++ order-inventory/src/main/java/com/spring/order_inventory/repository/CustomersRepository.java
@@ -3,5 +3,5 @@ package com.spring.order_inventory.repository;
 import com.spring.order_inventory.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomersRepository extends JpaRepository<Customer,Long> {
+public interface CustomersRepository extends JpaRepository<Customer,Integer> {
 }
