+++ order-inventory/src/main/java/com/spring/order_inventory/repository/ShipmentsRepository.java
@@ -3,5 +3,5 @@ package com.spring.order_inventory.repository;
 import com.spring.order_inventory.entity.Shipment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShipmentsRepository extends JpaRepository<Shipment,Long> {
+public interface ShipmentsRepository extends JpaRepository<Shipment,Integer> {
 }
